MAINT: Refine project structure
