+++ CMakeLists.txt
@@ -1,25 +1,15 @@
-# CMakeLists.txt
-add_definitions(-std=c++20)
-set(CMAKE_CXX_STANDARD 20)
-set(CMAKE_CXX_STANDARD_REQUIRED ON)
-cmake_minimum_required(VERSION 3.26)
+# CMakeList.txt : Top-level CMake project file, do global configuration
+# and include sub-projects here.
+cmake_minimum_required (VERSION 3.8)
+set (CMAKE_CXX_STANDARD 17)
 
-project(3pg)
+# Enable use of _ROOT variables
+if(POLICY CMP0074)
+  cmake_policy(SET CMP0074 NEW)
+endif()
 
-# gdal package should be installed using vcpkg, which creates a folder containing 
-# GDALConfig.cmake (which needs to be added to the environment variables)
-find_package(GDAL CONFIG REQUIRED)
+project ("3pg2")
 
-include_directories(include)
-
-set(SOURCE_FILES
-	3pg_sc/3pg_src/3pg.cpp
-	3pg_sc/3pg_src/util.cpp
-	3pg_sc/3pg_src/The_3PG_model.cpp
-	3pg_sc/3pg_src/Data_io.cpp
-	3pg_sc/3pg_src/MYDate.cpp
-    3pg_sc/3pg_src/GDALRasterImage.cpp
-)
-
-add_executable(${PROJECT_NAME} ${SOURCE_FILES})
-target_link_libraries(3pg PRIVATE GDAL::GDAL)
+# Include sub-projects.
+include_directories("include")
+add_subdirectory ("apps")
