+++ apps/MYDate.cpp
@@ -1,0 +1,29 @@
+// MYDate.cpp: implementation of the MYDate class.
+
+/*
+All source code remains the property and copyright of CSIRO. 
+
+CSIRO accepts no responsibility for the use of 3PG(S) or of the model 3-PG in
+the form supplied or as subsequently modified by third parties. CSIRO disclaims
+liability for all losses, damages and costs incurred by any person as a result
+of relying on this software. 
+Use of this software assumes agreement to this condition of use
+*/
+//
+//////////////////////////////////////////////////////////////////////
+
+#include "MYDate.h"
+
+//////////////////////////////////////////////////////////////////////
+// Construction/Destruction
+//////////////////////////////////////////////////////////////////////
+
+MYDate::MYDate()
+{
+
+}
+
+MYDate::~MYDate()
+{
+
+}
