+++ apps/The_3PG_Model.hpp
@@ -1,0 +1,4 @@
+// 3PG model run routine.
+#include "Data_io.hpp"
+
+void runTreeModel( MYDate spMinMY, MYDate spMaxMY, bool spatial, long cellIndex );
