+++ docs/legacy/Site.txt
@@ -1,0 +1,41 @@
+//Temporary Site definition file generated by Arc3PG
+"Model mode"  3PG
+"Output directory"  \3pg_example\output
+//Initial values
+"Latitude"      -18.5
+"Fertility rating"  0.8
+"Soil class"      3
+"Maximum ASW"    180
+"Minimum ASW"    0.0
+"Start age"      1
+"End Age"      7
+"Start month"    1
+"Year planted"    1993
+"WFi"        1
+"WRi"        0.8
+"WSi"        2
+"StemNoi"        1111
+"ASWi"        60
+
+//Satellite only data
+
+//Age Dependant Fertility Modifiers
+
+//Series Data
+"Tavg"     \3pg_example\climate\tavg01.flt     \3pg_example\climate\tavg02.flt     \3pg_example\climate\tavg03.flt     \3pg_example\climate\tavg04.flt     \3pg_example\climate\tavg05.flt     \3pg_example\climate\tavg06.flt     \3pg_example\climate\tavg07.flt     \3pg_example\climate\tavg08.flt     \3pg_example\climate\tavg09.flt     \3pg_example\climate\tavg10.flt     \3pg_example\climate\tavg11.flt     \3pg_example\climate\tavg12.flt
+"Rain"     \3pg_example\climate\rain01.flt     \3pg_example\climate\rain02.flt     \3pg_example\climate\rain03.flt     \3pg_example\climate\rain04.flt     \3pg_example\climate\rain05.flt     \3pg_example\climate\rain06.flt     \3pg_example\climate\rain07.flt     \3pg_example\climate\rain08.flt     \3pg_example\climate\rain09.flt     \3pg_example\climate\rain10.flt     \3pg_example\climate\rain11.flt     \3pg_example\climate\rain12.flt
+"Vpd"     \3pg_example\climate\vpd01.flt     \3pg_example\climate\vpd02.flt     \3pg_example\climate\vpd03.flt     \3pg_example\climate\vpd04.flt     \3pg_example\climate\vpd05.flt     \3pg_example\climate\vpd06.flt     \3pg_example\climate\vpd07.flt     \3pg_example\climate\vpd08.flt     \3pg_example\climate\vpd09.flt     \3pg_example\climate\vpd10.flt     \3pg_example\climate\vpd11.flt     \3pg_example\climate\vpd12.flt
+"solar radtn"     \3pg_example\climate\swrad01.flt     \3pg_example\climate\swrad02.flt     \3pg_example\climate\swrad03.flt     \3pg_example\climate\swrad04.flt     \3pg_example\climate\swrad05.flt     \3pg_example\climate\swrad06.flt     \3pg_example\climate\swrad07.flt     \3pg_example\climate\swrad08.flt     \3pg_example\climate\swrad09.flt     \3pg_example\climate\swrad10.flt     \3pg_example\climate\swrad11.flt     \3pg_example\climate\swrad12.flt
+"Frost Days"     0     0     0     0     0     0     0     0     0     0     0     0
+
+//Management Tables
+
+//Outputs
+"LAI"  LAI.flt  1994  1 month 1
+"StemNo"  StemNo.flt  1994  1 month 1
+"MAI"  MAI.flt  1994  1 month 1
+"average dbh"  avdbh.flt  1994  1 month 1
+"BasArea"  BasArea.flt  1994  1 month 1
+"StandVol"  StndVol.flt  1994  1 month 1
+"NPP"  NPP.flt  1994  1 month 1
+"ASW"  ASW.flt  1994  1 month 1
