+++ docs/legacy/Species.txt
@@ -1,0 +1,51 @@
+//Temporary Species definition file generated by Arc3PG
+"Foliage:stem partitioning ratio @ D=2 cm"        0.75
+"Foliage:stem partitioning ratio @ D=20 cm"        0.2
+"Constant in the stem mass v. diam. relationship"        0.02
+"Power in the stem mass v. diam. relationship"        3.2
+"Maximum fraction of NPP to roots"        0.6
+"Minimum fraction of NPP to roots"        0.1
+"Minimum temperature for growth"        8
+"Optimum temperature for growth"        25
+"Maximum temperature for growth"        31
+"Days production lost per frost day"        0.0
+"Moisture ratio deficit for fq = 0.5"        0.5
+"Power of moisture ratio deficit"        5
+"Value of 'm' when FR = 0"        0
+"Value of 'fNutr' when FR = 0"        0.5
+"Power of (1-FR) in 'fNutr'"        0.5
+"Maximum stand age used in age modifier"        10
+"Power of relative age in function for fAge"        4
+"Relative age to give fAge = 0.5"        0.9
+"Maximum litterfall rate"        0.12
+"Litterfall rate at t = 0"        0.002
+"Age at which litterfall rate has median value"        13
+"Average monthly root turnover rate"        0.03
+"Maximum canopy conductance"        0.018
+"LAI for maximum canopy conductance"        3
+"Defines stomatal response to VPD"        0.1
+"Canopy boundary layer conductance"        0.2
+"Max. stem mass per tree @ 1000 trees/hectare"        350
+"Power in self-thinning rule"        2
+"Fraction mean single-tree foliage biomass lost per dead tree"        0
+"Fraction mean single-tree root biomass lost per dead tree"        0.2
+"Fraction mean single-tree stem biomass lost per dead tree"        0.2
+"Specific leaf area at age 0"        10
+"Specific leaf area for mature leaves"        7
+"Age at which specific leaf area = (SLA0+SLA1)/2"        2.5
+"Extinction coefficient for absorption of PAR by canopy"        0.5
+"Age at canopy cover"        0
+"Maximum proportion of rainfall evaporated from canopy"        0.15
+"LAI for maximum rainfall interception"        3
+"Canopy quantum efficiency"        0.09
+"Ratio NPP/GPP"        0.4
+"Branch and bark fraction at age 0"        0.3
+"Branch and bark fraction for mature stands"        0.12
+"Age at which fracBB = (fracBB0+fracBB1)/2"        2
+"Minimum basic density - for young trees"        0.4
+"Maximum basic density - for older trees"        0.5
+"Age at which rho = (rhoMin+rhoMax)/2"        4
+"Intercept of net v. solar radiation relationship"        -20
+"Slope of net v. solar radiation relationship"        0.8
+"Molecular weight of dry matter"        24
+"Conversion of solar radiation to PAR"        2.3
