+++ docs/legacy/run.bat
@@ -1,0 +1,1 @@
+3pg.exe -s site.txt -d species.txt
