--- include/The_3PG_Model.hpp
@@ -1,4 +1,0 @@
-// 3PG model run routine.
-#include "Data_io.hpp"
-
-void runTreeModel( MYDate spMinMY, MYDate spMaxMY, bool spatial, long cellIndex );
