--- run.bat
@@ -1,1 +1,0 @@
-3pg.exe -s site.txt -d species.txt
